# API/src/main/resources/static/swagger.yml
openapi: 3.0.0
info:
  title: PIR RDF plugin project API
  description: API documentation for the PIR RDF plugin project.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
...
    Annotation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        annotationType:
          type: string
        pageUrl:
          type: string
        turtlePath:
          type: string
        item:
          $ref: "#/components/schemas/Item"
        concepts:
          type: array
          items:
            $ref: "#/components/schemas/Concept"
        campaign:
          type: object
          properties:
            name:
              type: string
        creator:
          type: object
          properties:
            email:
              type: string
    Item:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
    Concept:
      type: object
      properties:
        value:
          type: string
        uri:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/user:
    get:
      tags:
        - User
      summary: Get all users
      operationId: getUsers
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden - user not authorized
          content:
            text/plain:
              schema:
                type: string
                example: Permission denied. Disappear !
...
  /api/annotation:
    post:
      tags:
        - RDF
      summary: Create a new annotation
      operationId: createAnnotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                annotationType:
                  type: string
                item:
                  type: object
                  properties:
                    type:
                      type: string
                    value:
                      type: string
                pageUrl:
                  type: string
                turtlePath:
                  type: string
                concepts:
                  type: array
                  items:
                    type: object
                    properties:
                      value:
                        type: string
                      uri:
                        type: string
                campaign:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
              required:
                - description
                - annotationType
                - item
                - pageUrl
                - turtlePath
                - concepts
                - campaign
      responses:
        "201":
          description: Annotation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "401":
          description: Unauthorized - account issue
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized
        "404":
          description: Resource not found
          content:
            text/plain:
              schema:
                type: string
                example: CampaignNotFound | Incorrect email.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
    # patch:
    #   tags:
    #     - RDF
    #   summary: Update an annotation
    #   operationId: updateAnnotation
    #   parameters:
    #     - name: id
    #       in: query
    #       required: true
    #       schema:
    #         type: integer
    #         format: int64
    #     - name: name
    #       in: query
    #       required: true
    #       schema:
    #         type: string
    #   responses:
    #     "200":
    #       description: Annotation updated successfully
    #       content:
    #         application/json:
    #           schema:
    #             $ref: "#/components/schemas/Annotation"
    #     "401":
    #       description: Unauthorized - account issue
    #       content:
    #         text/plain:
    #           schema:
    #             type: string
    #             example: Unauthorized
    #     "404":
    #       description: Annotation not found
    #       content:
    #         text/plain:
    #           schema:
    #             type: string
    #             example: Annotation not found.
    #     "500":
    #       description: Internal server error
    #       content:
    #         text/plain:
    #           schema:
    #             type: string
    #             example: Oops, error. Please contact an administrator.
  /api/annotation/user:
    get:
      tags:
        - RDF
      summary: Get annotations for current user
      operationId: getAnnotationsByUser
      responses:
        "200":
          description: Annotations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Annotation"
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/plugin:
    get:
      tags:
        - Plugin
      summary: Download plugin
      operationId: downloadPlugin
      responses:
        "200":
          description: Plugin downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.